feat: balance operations, behind the accountant right
There was no way to put money into an account or take it out. Every account
funded itself once from its group's opening deposit and that was all.

The server side follows the order surface: a shared core in balance.go, an
http adapter and a ws adapter, all reaching the engine on
system.s<shard>.balance. Five routes: a general one that takes any money
action, and deposit, withdrawal, credit and correction for the common cases.
A client can also do it over the socket. Every one of them is gated on the
accountant right and on the caller's group masks, and every one writes a
journal line naming the actor, the account and the amount.

The engine applies it. Credit and bonus move credit rather than balance,
because credit counts toward margin but is not the client's money. Only a
correction may leave an account short; a withdrawal or a charge that would
take the balance below zero is refused. Every operation writes a deal with
its own action, so the ledger explains the balance, and the account is added
up again afterwards so equity and free margin follow.

Also fixed a comparison in the merged datafeed code that narrowed a count to
int32 before checking a bound, which could have wrapped a large value into a
passing one.

+++ hst-server/internal/server/v1/datafeeds.go
@@ -1190,7 +1190,8 @@ func (s *HttpServer) DeleteDatafeedParam(c *fiber.Ctx) error {
 var errParamPriorityOutOfRange = errors.New("priority out of range")
 
 func validateDatafeedParamPriority(target int32, count int) error {
-	if target < 0 || target >= int32(count) {
+	// compare as int64 so a count larger than an int32 cannot wrap into a passing value
+	if target < 0 || int64(target) >= int64(count) {
 		return errParamPriorityOutOfRange
 	}
 	return nil
